Add TargetAllyWeaponSpell and integrate usage
-  reusable TargetAllyWeaponSpell helper and Skilltarget enum value to centralize ally selection for weapon-spell casts.

- Py4GWCoreLib/Builds/Skills/ritualist/RestorationMagic.py: refactored Xinraes_Weapon to use TargetAllyWeaponSpell instead of duplicating selection logic.
